Move apng tests into test_file_apng.py
